requirements.txt
openai==2.26.0
python-dotenv==1.2.2
